entropy-engine/pyproject.toml
[tool.poetry]
name = "entropy-engine"
version = "0.1.0"
description = "SRE Masterclass Entropy Engine"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
structlog = "^25.4.0"
pyyaml = "^6.0.2"
docker = "^7.1.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
